moved coordinate header to data class directory
