+++ src/dataClasses/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(artss
+        PRIVATE
+        ${CMAKE_CURRENT_LIST_DIR}/Coordinate.h
+)
