+++ src/dataClasses/Coordinate.h
@@ -13,7 +13,7 @@
 #include <vector>
 #include "../utility/GlobalMacrosTypes.h"
 #ifndef BENCHMARKING
-#include <fmt/format.h>
+#include "fmt/format.h"
 #endif
 
 inline static const std::vector<std::string> axis_names = {"X", "Y", "Z"};
